move Audience class to a separate file
